Added function overloads for `write` and `resize` to accept x and y together as `size_t[2]`.  Also added a `size` method for `w` and `h`.
Using an array for pos and size is not entirly common in C++ due to arrays being a pain to use.
But in higher level languages like python (particularly pygame) this is used.
And with D arrays are easier to use than in C++ so some coders (like myself) may prefer using an array.